Implemented solution that loops over the range, calculating the hamming weight each time using this technique
http://graphics.stanford.edu/~seander/bithacks.html#CountBitsSetParallel

Given we are calculating the hamming weight in sequence I assume there
is a relationship between i and i-1 so some there must be a more
efficient solution using a dynamic programming style solution.

The other one to consider would be building a lookup table if the
range of n is known and not too large. Even if this is not reasonable
for the larger values of N, precalculating the first x in a lookup table
would most likely pay off as we are always starting from 0.